Refact
Refact renderdevice to support multiple API

+++ glbcodebase/graphicslab_gl/glb/glbmodel.cpp
@@ -236,7 +236,7 @@ void ModelMgrImp::Initialize() {
 }
 
 void ModelMgrImp::Destroy() {
-    for (int32_t i = 0; i < m_ModelDataBase.size(); i++) {
+    for (int32_t i = 0; i < static_cast<int32_t>(m_ModelDataBase.size()); i++) {
         GLB_SAFE_DELETE(m_ModelDataBase[i]);
     }
     m_ModelDataBase.clear();
